feat(golang): calculate area

+++ lambda/Shape/pkg/models/request.go
@@ -1,8 +1,8 @@
 package models
 
 type Request struct {
-	Creator   string `json:"creador"`
-	ShapeType string `json:"tipo"`
-	A         int64  `json:"a"`
-	B         int64  `json:"b"`
+	Creator   string  `json:"creador"`
+	ShapeType string  `json:"tipo"`
+	A         float64 `json:"a"`
+	B         float64 `json:"b"`
 }
